Edited localized.properties to push LocalizedMutation to the front of the list of observers - it does not appear to have had a substantive effect.
Edited WorldCrawl in an attempt to discover why several locations tend to be missing from speciesConcentrationMap when getAmp is called - have tracked it back to the fact that migration is performed _before_ mutation, but _after_ generationData.  Played around with a few methods of alleviating this, but have run into other confusing problems.  "world" becomes unset very easily.  Details on how to elicit this behavior are in comments, just before the declaration of "generateConcentrationMap".

Additionally altered GARun to label the generationData call before _anything_ happens as generation -1.  This was to avoid the old condition of there simply being two generation 0s.

git-svn-id: http://cscirepo.morris.umn.edu/svn/research/LandscapeEC@8303 e5a9bd0e-96e4-4e5d-b9da-31dc9101860d